Actualización de funcionamiento y mejora de panel de director y la comunicación con la base de datos, y eliminación de dos tablas de la bd (roles, usuarios)

+++ backend/backend/src/main/java/appesperanzaviva/backend/repository/SolicitudRepository.java
@@ -15,5 +15,5 @@ public interface SolicitudRepository extends JpaRepository<Solicitud, Long> {
     long countByEstado(String estado);
 
     // 🔹 NUEVO: Buscar solicitudes asignadas a un conciliador especifico
-    java.util.List<Solicitud> findByConciliadorId(Long conciliadorId);
+    java.util.List<Solicitud> findByConciliadorId(Integer conciliadorId);
 }
